+++ org.adempiere.ui.zk/WEB-INF/src/org/adempiere/webui/info/InfoInvoiceWindow.java
@@ -76,7 +76,7 @@ public class InfoInvoiceWindow extends InfoWindow {
 		super.saveSelectionDetail();
 
         //  publish for Callout to read
-        Integer ID = getSelectedRowKey();
+        Integer ID = (Integer) getSelectedRowKey();
         Env.setContext(Env.getCtx(), p_WindowNo, Env.TAB_INFO, "C_Invoice_ID", ID == null ? "0" : ID.toString());
 	}
