+++ org.adempiere.ui.zk/WEB-INF/src/org/adempiere/webui/panel/InfoProductPanel.java
@@ -690,7 +690,7 @@ public class InfoProductPanel extends InfoPanel implements EventListener<Event>
 			rs = null; pstmt = null;
 		}
 
-		m_M_Product_ID = getSelectedRowKey();
+		m_M_Product_ID = (Integer) getSelectedRowKey();
 		sql = "SELECT DocumentNote FROM M_Product WHERE M_Product_ID=?";
 		fieldDescription.setText(DB.getSQLValueString(null, sql, m_M_Product_ID));
 
@@ -1170,7 +1170,7 @@ public class InfoProductPanel extends InfoPanel implements EventListener<Event>
 	protected void showHistory()
 	{
 		log.info("");
-		Integer M_Product_ID = getSelectedRowKey();
+		Integer M_Product_ID = (Integer) getSelectedRowKey();
 		if (M_Product_ID == null)
 			return;
 		int M_Warehouse_ID = 0;
@@ -1204,7 +1204,7 @@ public class InfoProductPanel extends InfoPanel implements EventListener<Event>
 	public void zoom()
 	{
 		log.info("");
-		Integer M_Product_ID = getSelectedRowKey();
+		Integer M_Product_ID = (Integer) getSelectedRowKey();
 		if (M_Product_ID == null)
 			return;
 
@@ -1248,7 +1248,7 @@ public class InfoProductPanel extends InfoPanel implements EventListener<Event>
 	protected void saveSelectionDetail()
 	{
 		//  publish for Callout to read
-		Integer ID = getSelectedRowKey();
+		Integer ID = (Integer) getSelectedRowKey();
 		Env.setContext(Env.getCtx(), p_WindowNo, Env.TAB_INFO, "M_Product_ID", ID == null ? "0" : ID.toString());
 		ListItem pickPL = (ListItem)pickPriceList.getSelectedItem();
 		if (pickPL!=null)
@@ -1399,7 +1399,7 @@ public class InfoProductPanel extends InfoPanel implements EventListener<Event>
     	int row = contentPanel != null ? contentPanel.getSelectedRow() : -1;
 		if (component.equals(m_PAttributeButton) && row != -1)
 		{
-			Integer productInteger = getSelectedRowKey();
+			Integer productInteger = (Integer) getSelectedRowKey();
 			String productName = (String)contentPanel.getValueAt(row, INDEX_NAME);
 
 			ListItem warehouse = pickWarehouse.getSelectedItem();
